Add extensibility to the Block Bindings UI in the editor
