+++ packages/editor/src/components/block-bindings/style.scss
@@ -10,7 +10,7 @@
 		border-bottom: 1px solid #0002;
 	}
 
-	.block-bindings-metadata-source-ui {
+	.block-bindings-fields-list-ui {
 		padding: 12px;
 		li {
 			margin: 20px 8px;
